+++ includes/mb/class.pitemp.inc.php
@@ -14,7 +14,7 @@ class PiTemp extends Sensors
 {
     private function _temperature()
     {
-        if (file_exists('/sys/devices/platform/sunxi-i2c.0/i2c-0/0-0034/temp1_input')) { //Banana Pi
+        if (file_exists('/sys/devices/platform/sunxi-i2c.0/i2c-0/0-0034/temp1_input')) { // Banana Pi
             $temp = file_get_contents('/sys/devices/platform/sunxi-i2c.0/i2c-0/0-0034/temp1_input');
             $temp_max = -1;
         } else {
@@ -30,8 +30,32 @@ class PiTemp extends Sensors
         $this->mbinfo->setMbTemp($dev);
     }
 
+    private function _voltage()
+    {
+        if (file_exists('/sys/devices/platform/sunxi-i2c.0/i2c-0/0-0034/axp20-supplyer.28/power_supply/ac/voltage_now')) { // Banana Pi
+            $volt = file_get_contents('/sys/devices/platform/sunxi-i2c.0/i2c-0/0-0034/axp20-supplyer.28/power_supply/ac/voltage_now');
+            $dev = new SensorDevice();
+            $dev->setName("Voltage 1");
+            $dev->setValue($volt / 1000000);
+            $this->mbinfo->setMbVolt($dev);
+        }
+    }
+
+    private function _current()
+    {
+        if (file_exists('/sys/devices/platform/sunxi-i2c.0/i2c-0/0-0034/axp20-supplyer.28/power_supply/ac/current_now')) { // Banana Pi
+            $current = file_get_contents('/sys/devices/platform/sunxi-i2c.0/i2c-0/0-0034/axp20-supplyer.28/power_supply/ac/current_now');
+            $dev = new SensorDevice();
+            $dev->setName("Current 1");
+            $dev->setValue($current / 1000000);
+            $this->mbinfo->setMbCurrent($dev);
+        }
+    }
+
     public function build()
     {
         $this->_temperature();
+        $this->_voltage();
+        $this->_current();
     }
 }