+++ CHANGELOG.md
@@ -10,7 +10,7 @@ phpSysInfo 3.2.0
  - [ADD] Catalan Translation ca.xml
  - [ADD] Mer to detected distros (core of Sailfish OS)
  - [ADD] ThermalZone sensor support for Linux and Android
- - [ADD] Tempsensor for Banana Pi 
+ - [ADD] Temperature, voltage and current sensor for Banana Pi
  
  - [UPD] German Translation de.xml